#18 - [REFACTOR] 기본 생성자에서 체이닝 방식으로 변경
